modified 12

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
  */
 void print_number(int n)
 {
-	int a, b, c, d, e;
+	long int a, b, c, d, e;
 
 	b = n;
 	e = a = 1;
